LPD-90764 Extract fetchAllPages helper from category-tree hook
