+++ modules/apps/frontend-js/frontend-js-item-selector-web/src/main/resources/META-INF/resources/item_selector/useTaxonomyCategoryTreeNodes.ts
@@ -3,17 +3,15 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import {DEFAULT_FETCH_HEADERS} from '@liferay/frontend-data-set-web';
-import {fetch} from 'frontend-js-web';
 import {useEffect, useMemo, useState} from 'react';
+
+import fetchAllPages from './fetchAllPages';
 
 // The vocabulary-scoped endpoint is used (not /sites/{id}/taxonomy-categories)
 // because the site path filters out categories of global vocabularies, while
 // this path returns the full set regardless of the requesting site.
 
 const HEADLESS_TAXONOMY_BASE = '/o/headless-admin-taxonomy/v1.0';
-
-const PAGE_SIZE = 100;
 
 export interface ITaxonomyCategory {
 	availableLanguages?: string[];
@@ -39,52 +37,6 @@ export interface ITaxonomyCategoryTreeNode {
 	name: string;
 	raw: ITaxonomyCategory;
 	vocabularyId: string;
-}
-
-interface IPage<T> {
-	items?: T[];
-	lastPage?: number;
-	page?: number;
-	totalCount?: number;
-}
-
-async function fetchAllPages<T>(
-	baseURL: string,
-	signal: AbortSignal
-): Promise<T[]> {
-	const items: T[] = [];
-
-	let page = 1;
-	let lastPage = 1;
-
-	do {
-		const url = new URL(baseURL, window.location.origin);
-
-		url.searchParams.set('page', String(page));
-		url.searchParams.set('pageSize', String(PAGE_SIZE));
-
-		const response = await fetch(url.toString(), {
-			headers: DEFAULT_FETCH_HEADERS,
-			signal,
-		});
-
-		if (!response.ok) {
-			throw new Error(
-				`Request to ${url.pathname} failed: ${response.status}`
-			);
-		}
-
-		const json = (await response.json()) as IPage<T>;
-
-		if (Array.isArray(json.items)) {
-			items.push(...json.items);
-		}
-
-		lastPage = json.lastPage ?? 1;
-		page += 1;
-	} while (page <= lastPage);
-
-	return items;
 }
 
 function buildTree(
